Uses one event handler and data to remember associated hint text
as as opposed to an individual event with a closure for each item. 
This is similar to the previous listing, except that i have used data instead of a closure. The key benefit is that I have moved the event handling setup out of the initial each loop, so that there is only one event handler function, as opposed to each item having its own customized handler function.