used .css, it ended up being less glitchy
The real problem with sliding up and down in jQuery is that the code won't know the actual size of the element until it is fully expanded. Often the animation will incorrectly estimate the actual size, resizing based on an incorrect value, and then have to catch up at the end of the effect. If we let jQuery know what the actual size is, it can perform a better interpolation.

This solution uses visibility: hidden instead of display: none at the start, in order that the sliding element have its natural size. We then measure that size, and explicitly set it with the css function, while at the same time reversing the visibility and display to hidden and none respectively. This could result in a slight flicker as the page loads, but, in practice, the effect seems negligible. Having the correct final size will enable jQuery to better interpolate the height as it slides the element.

We have also used a linear easing, instead of the default swing, which improves the effect even more.